+++ src/SMAPI/Events/SavePreloadedEventArgs.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace StardewModdingAPI.Events
+{
+    /// <summary>Event arguments for an <see cref="ISpecialisedEvents.SavePreloaded"/> event.</summary>
+    public class SavePreloadedEventArgs : EventArgs { }
+}