+++ src/SMAPI/Framework/Events/ModSpecialisedEvents.cs
@@ -9,6 +9,13 @@ namespace StardewModdingAPI.Framework.Events
         /*********
         ** Accessors
         *********/
+        /// <summary>Raised immediately after the player loads a save slot, but before the world is fully initialised. The save and game data are available at this point, but some in-game content (like location maps) haven't been initialised yet.</summary>
+        public event EventHandler<SavePreloadedEventArgs> SavePreloaded
+        {
+            add => this.EventManager.SavePreloaded.Add(value);
+            remove => this.EventManager.SavePreloaded.Remove(value);
+        }
+
         /// <summary>Raised before the game state is updated (≈60 times per second), regardless of normal SMAPI validation. This event is not thread-safe and may be invoked while game logic is running asynchronously. Changes to game state in this method may crash the game or corrupt an in-progress save. Do not use this event unless you're fully aware of the context in which your code will be run. Mods using this event will trigger a stability warning in the SMAPI console.</summary>
         public event EventHandler<UnvalidatedUpdateTickingEventArgs> UnvalidatedUpdateTicking
         {